Add files via upload
Coding a 3-6V motor to operate in conjunction with a joystick module. The joystick position in the x axis dictates both the direction of rotation and it's speed. L293D chip is used and a power supply module. If another motor is used you should check the chip matches the motor's current requirements or the motor may draw too much current.