Add files via upload
Serial to parallel shift register using 74HC595 IC chip. The code as it is makes the four LEDs on the left turn on while the right four are on and vice versa in a loop. This can be easily manipulated into a binary counter by using an increment command inside the loop.